requirements.txt
aiofiles==22.1.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
choicelib==0.1.5
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
iso8601==1.1.0
loguru==0.7.2
multidict==6.0.4
pydantic==1.10.13
pypika-tortoise==0.1.6
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
win32-setctime==1.1.0
yarl==1.9.4
